Fix a potential SIGSEGV in gtpv2 C code for > 32 bit allocations
Summary:
NwHandleT holds pointers to generic handles structs in the gtpv2c code.
Specifically this holds the reference to NwGtpv2cTmrMinHeapT struct which
contains pointer to the timeout timers heap.
In certain situations like in NwGtpv2c.c:107
*thiz = (NwGtpv2cTmrMinHeapT *) malloc (sizeof (NwGtpv2cTmrMinHeapT));
the address of thiz may > 2^32 causing the stored pointer to be invalid, and
thus causing a subsequent SIGSEGV when a new timer is inserted into the heap.
Fix this by making the handle large enough

Test Plan:
Reproduced issue after core by forcing malloc allocations at a large enough address
and verified it goes away after the fix.

+++ SRC/GTPV2-C/nwgtpv2c-0.11/shared/NwTypes.h
@@ -55,7 +55,7 @@ extern "C" {
 typedef signed int              NwFdT;          /**< Use this for file descriptor */
 
 typedef uintptr_t               NwPtrT;         /**< Use this for generic pointers      */
-typedef unsigned int            NwHandleT;      /**< Use this for generic handles       */
+typedef uintptr_t               NwHandleT;      /**< Use this for generic handles       */
 
 
 #ifdef __cplusplus
